+++ alioth/src/virtio/dev/blk.rs
@@ -86,7 +86,7 @@ const SECTOR_SIZE: usize = 1 << 9;
 
 bitflags! {
     #[derive(Default, Debug, Clone, Copy, PartialEq, Eq, Hash)]
-    pub struct BlockFeature: u64 {
+    pub struct BlockFeature: u128 {
         const SIZE_MAX = 1 << 1;
         const SEG_MAX = 1 << 2;
         const GEOMETRY = 1 << 4;
@@ -293,7 +293,7 @@ impl Virtio for Block {
         self.config.clone()
     }
 
-    fn feature(&self) -> u64 {
+    fn feature(&self) -> u128 {
         self.feature.bits() | FEATURE_BUILT_IN
     }
 
@@ -320,7 +320,7 @@ impl VirtioMio for Block {
 
     fn activate<'a, 'm, Q, S, E>(
         &mut self,
-        _feature: u64,
+        _feature: u128,
         _active_mio: &mut ActiveMio<'a, 'm, Q, S, E>,
     ) -> Result<()>
     where
@@ -415,7 +415,7 @@ impl VirtioMio for Block {
 impl VirtioIoUring for Block {
     fn activate<'a, 'm, Q, S, E>(
         &mut self,
-        _feature: u64,
+        _feature: u128,
         _ring: &mut ActiveIoUring<'a, 'm, Q, S, E>,
     ) -> Result<()>
     where
